Add 'reverse' setting.
This setting indicate whether coordinates in the field value need to be
reversed before using in GeoJSON point object. GeoJSON expects coordinates to
be recorded in LngLat order. If this is not the case for the field value
content, specify 'reverse: true' in autoform field settings, and filed value
will be reversed prior to used in GeoJSON point object:

location: {
  type: twodsphereSchema,
  index: "2dsphere",
  autoform: {
    type: 'geojson-point',
    afFieldInput: {
      reverse: true
    }
  }
},